ajout service Distributeur + modification typage int

+++ lille1/src/main/java/fil/tiir/fakedistrib/entity/Distributeur.java
@@ -26,7 +26,7 @@ public class Distributeur {
 		return this.montant;
 	}
 
-	public void diminuerMontant(double montant) {
+	public void diminuerMontant(int montant) {
 		this.montant -= montant;
 	}
